+++ src/main/java/netcrafts/detective/query/ChunkStatusUtil.java
@@ -3,6 +3,7 @@ package netcrafts.detective.query;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
 
+@SuppressWarnings("null")
 public class ChunkStatusUtil {
 
     /**